WIP: FABRIK

+++ CS_6610/Animation/FABRIK.h
@@ -30,7 +30,7 @@ void SolveForward(const FABRIKParams& i_params);
 void SolveBackward(const FABRIKParams& i_params);
 void UpdateRotations(const FABRIKParams& i_params);
 
-float CalculateChainLength(const Skeleton* i_skeleton, uint16_t i_rootJointIndex, uint16_t i_endJointIndex);
+float CalculateChainLength(const Skeleton* i_skeleton, uint8_t i_rootJointIndex, uint8_t i_endJointIndex);
 void PrintSolvedJoints(const FABRIKParams& i_params);
 
 } // namespace animation
